Change wine category form from products to administration app
